Major feature update: Tailwind, auth, forum, profiles, admin, WebSocket
- Switch to Tailwind CSS via CDN with Phosphor Icons
- Auth: login, register, JWT cookies, nav dropdown with profile/dashboard/settings/admin
- Forum: categories, threads, posts, replies with notifications
- Profiles: public profiles with bio, location, gender, banner/profile colors, badges, follow system
- Notifications: bell with unread count, WebSocket live updates
- WebSocket: real-time events for notifications, credit updates, new posts
- Admin panel: users, assets, categories, disputes, settings, docs, forum categories, badges, roles, bans, mod notes
- Role system: granular JSONB permissions, multiple roles per user, staff flags
- Ban system: timed or permanent bans, ban enforcement on forum posts
- Mod notes: staff-only notes on user profiles
- Profile customization: colors, bio, location, gender, website
- Settings page: edit profile, change password, communication preferences
- Download page: Windows/macOS/Linux with platform icons
- Fix axum 0.7 route params (:param not {param})
- Fix COUNT/SUM/MAX SQL casts for sqlx compatibility
- Access token extended to 30 days
- Explicit SSR route registration (no fallback_service conflicts)
- 6 new DB migrations (003-006)

+++ crates/models/src/article.rs
@@ -132,7 +132,7 @@ impl Article {
 
         let total: (i64,) = sqlx::query_as(
             r#"
-            SELECT COUNT(*) FROM articles a
+            SELECT COUNT(*)::bigint FROM articles a
             WHERE a.published = true
               AND ($1::text IS NULL OR $1 = ANY(a.tags))
             "#,
